fix: 修复demo json下拉不可编辑问题；使用useFullscreen hook
