+++ src/utils/DragUtil.ts
@@ -1,8 +1,6 @@
 import { useDraggable } from "@vueuse/core"
 import { Ref, StyleValue, watchEffect } from "vue"
 import { DragableConf } from "@/models/DraggableInterface"
-import { _isSafari } from "@/Util"
-import { useFullscreenListener } from "./ListenerUtil"
 
 function getLengthStyle(length?: number | string): string | undefined {
   if (typeof length === "number") {
@@ -30,83 +28,6 @@ export function getDragStyle(dragConf: DragableConf): StyleValue {
     top: initY,
     borderRadius: "15px",
   }
-}
-
-/**
- * 处理全屏事件 - 不再支持IE11
- *
- * @param fullscreen
- * @param fullArea
- */
-export function useToggleFullscreen(fullscreen: Ref<boolean>, fullArea: Ref<HTMLDivElement | undefined>) {
-  useFullscreenLifecycle(fullscreen, fullArea)
-  return () => {
-    if (fullscreen.value) {
-      document.exitFullscreen()
-    } else {
-      fullArea.value?.requestFullscreen()
-    }
-    fullscreen.value = !fullscreen.value
-  }
-}
-
-type SafariStyleCache = {
-  position: string
-  width: string
-  height: string
-  left: string
-  top: string
-}
-
-const defaultFullScreenStyle = {
-  position: "fixed",
-  width: "100%",
-  height: "100%",
-  left: "0",
-  top: "0",
-}
-
-export function useFullscreenLifecycle(fullscreen: Ref<boolean>, fullArea: Ref<HTMLDivElement | undefined>) {
-  // TODO 考虑用成熟的全屏库实现, 例如 vueuse
-  let safariStyleCache: SafariStyleCache = { ...defaultFullScreenStyle }
-  useFullscreenListener(() => {
-    const isFullScreen = document.fullscreenElement
-    if (isFullScreen) {
-      const container = fullArea.value
-      if (!container) {
-        return
-      }
-      //  进入全屏
-      if (_isSafari()) {
-        safariStyleCache = {
-          position: container.style.position,
-          width: container.style.width,
-          height: container.style.height,
-          left: container.style.left,
-          top: container.style.top,
-        }
-        container.style.position = defaultFullScreenStyle.position
-        container.style.width = defaultFullScreenStyle.width
-        container.style.height = defaultFullScreenStyle.height
-        container.style.left = defaultFullScreenStyle.left
-        container.style.top = defaultFullScreenStyle.top
-      }
-    } else {
-      //  退出全屏
-      fullscreen.value = false
-      const container = fullArea.value
-      if (!container) {
-        return
-      }
-      if (_isSafari()) {
-        container.style.position = safariStyleCache.position
-        container.style.width = safariStyleCache.width
-        container.style.height = safariStyleCache.height
-        container.style.left = safariStyleCache.left
-        container.style.top = safariStyleCache.top
-      }
-    }
-  })
 }
 
 export function useDrag(
